+++ src/features/login/ui/index.ts
@@ -1,1 +1,1 @@
-export * from './login-container';
+export * from './LoginForm';
