+++ src/widgets/container/index.ts
@@ -1,1 +1,3 @@
+export * from './AuthContainer';
+export * from './AuthLinkContainer';
 export * from './logo-container';
